refactor(lazy-chat): move crisp to provider folder
